Добавляет сэмпл-карточку на главную n-раз
